+++ XRTK-Core/Assets/XRTK.Seed/PackageManifest.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 510fe338c9d74c459b402b2fb28b14f0
+timeCreated: 1588362284
