+++ XRTK-Core/Assets/XRTK.Seed/ScopedRegistry.cs
@@ -1,0 +1,20 @@
+﻿// Copyright (c) XRTK. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using Newtonsoft.Json;
+using System.Collections.Generic;
+
+namespace XRTK.Seed
+{
+    internal class ScopedRegistry
+    {
+        [JsonProperty("name")]
+        public string Name { get; set; }
+
+        [JsonProperty("url")]
+        public string Url { get; set; }
+
+        [JsonProperty("scopes")]
+        public List<string> Scopes { get; set; }
+    }
+}
